Reorganized components and implemented delete app resource functionality
